README.rst: Fix path for badge images
These badges are not committed to `bst-1.2` branch and hence are not
available on docs.buildstream.build website.

As it was mentioned in the issue comments, the correct place to obtain
these badges is always from the `master` branch, which is what the
buildstream.gitlab.io/buildstream website corresponds to.

Fixes #1056.