loader.py: Check source cache for subproject fetching
If the junction source consistency is cached but the junction is not in
the source cache, also use Stream.fetch_subprojects() to import sources
into source cache instead of doing this implicitly in
Element._stage_sources_at(). This is in preparation for buildbox-casd,
which will disallow local CAS cache writes from the main process (gRPC).