_context:extract _messenger.Messenger from Context
Separate out the responsibility of messaging from the rest of the
Context object.

This change removes some member variables and some code from member
functions, but keeps the existing member functions. In later work we'll
want things to use the Messenger directly, so we can remove some member
functions too, and de-couple further.

Update some comments to refer to the "Messenger API" instead of the
"Context API" in the copied code.

Fixup the arg documentation to timed_activity() while we're there.