_yaml: Refer new synthetic nodes to their parents
This changes how we were handling synthetic nodes. Previously, we would
have them as coming from a 'SYNTHETIC_FILE'. This is not correct when
we need the position in the real file when adding new information to a
file we will end up writing to disk (like project_ref).

This is a prerequisite for further work on cleaning how it is done.