_messenger: clarify silent_nested behaviour
Change the argument name on the 'silence' context manager, to avoid the
interpretation that non-nested messages may be silenced, or that it can
'un-silence' (it can't).

Be more specific about what 'important' means w.r.t. messages.

Require 'actually_silence' to be specified as a keyword, so that
callsites are easier to understand.