utils.py: Ensure dest copy directory is writable
In the case we are copying something to a read-only directory,
`_copy_directories` fails. Copying should work regardless.
Now when we are copying something to a read-only directory, we
reset the permissions of the directory to writable, and then revert to
the directory's original permissions after copying is completed.

An example where this is nessesary is when extracting Docker layers.
If one layer adds a read-only directory `/a` and the following layers adds
`/a/foo`, we must be able to copy `foo` into the otherwise read-only directory,
whilst still ensuring `/a` is read-only once the layer is staged.