src/buildstream/testing/_utils/site.py: Disable bwrap when not using sanbox
In order to test locally the code for chroot, one needs to define
`BST_FORCE_BACKEND`. But many tests use HAVE_BWRAP only to test
whether a test is made for bwrap, even if BuildStream will ignore
bwrap. To simplify, we disable bwrap when not using Linux backend so
one does not have to uninstall bwrap.