_frontend: Added toplevel `--max-jobs` configuration
While this is currently only relevant for `bst build`, it is
a current implementation detail that user configuration which
gets overridden by command line options must use toplevel
options.

This patch allows invocations such as the following to override
the max-jobs:

    bst --max-jobs 16 build target.bst

This also updates the completions test to expect the new toplevel option.

This is a part of #1033