tests/workspaces: Test edge case when linking files to RO directory
When we open a workspace, we are linking the sources to the directory
specified. As such testing that we are able to open a workspace on a
read-only directory should show that we are able to link files to a RO
directory when we are staging complex sources (eg. Docker source).