_loader: Detect and reject duplicated dependencies
When loading an element, if it has duplicated dependencies of any kind
then we reject the element with a LoadError(DUPLICATE_DEPENDENCY).  This
means that an element `foo.bst` can only appear once in any given dependency
kind, allowing it to be both `runtime` and `build`, but not either of those
twice, nor either plus `all`, nor `all` twice.

Signed-off-by: Daniel Silverstone <daniel.silverstone@codethink.co.uk>