element.py: Don't reset a failed (but cached) workspaced Element
A workspaced element can fail to build, if this happens, we
cache the result as a failure. This patch ensures that we
do not reset the cache data if a workspace has failed to build.

An integration test has also been added which ensures that we
are able to get the log of a failed workspace build.

Closes #1096