tests: Use local plugins in variable tests
Currently we use plugins sourced from bst-plugins-experimental for the
overrides tests, however this should be unnecessary and adds an external
dependency on bst-plugins-experimental. This change uses a local plugins
repository storing copies of the plugins from bst-plugins-experimental.
These do not even need to be kept up to date, as the tests are simply
for checking if variables are loaded and overwritten correctly.