tests/internals/storage_vdir_import.py: Reduce number of test cases
In changing the tests to use in_subprocess, they were modified to do
parametrization over both variables instead of looping over one in the
test itself. The purpose of the original code was to reduce the number
of test cases printed in the output. I have hence attempted to remove
superfluous test cases.