sandbox/_mounter: Remove default mutable arguments stderr/out
stderr and stdout were passed as default arguments and would therefore
retain the first value they had when the module was imported, which
means they wouldn't get overriden by the pytest capture of stderr/out.

This also meant that depending how the mounter was imported, if the
stdout/err was patched at that moment, and the file closed, we would
get an error.

The bug can be reproduced by running:

tox -e pyXX -- tests/integration/cmake.py

On master, and seeing that it is now fixed.