_stream.py: Remove separate handling of ArtifactElements
ArtifactElement inherits Element, both have an Artifact
object as a member, thus we should not need to handle these
separately.

This change has resulted in introducing a configure_sandbox()
method in ArtifactElement. The method is similar to
BuildElement.configure_sandbox() but does not configure the
sandbox to actually be used for building.