scheduler.py: Notification for interactive failure retry
Add a notifcation for RETRY. This moves the retry handling into
scheduler, which will be running in the process which has been
suspended for interactivity and as such will be able to load the
relevant Element. Note a failed job via the scheduler should never
not have a related queue, so the try except when matching the queue
via the action name should not be needed.