_project.py: Don't only load cached artifacts
Previously, before loading artifact refs (ArtifactElements) we were
checking that they were cached before loading them. This obviously becomes
a problem when we want to try and pull an artifact. This patch removes
the check for cached artifacts as we should still be able to contruct
an ArtifactElement regardless of whether it is cached or not.