tests: Add a basic plugin to test variables
As we're moving all plugins out of this repo, but we still should
probably test that variables can be loaded from defaults and overrided
correctly, I've added a very basic test plugin to handle this. It's
essentially just a BuildElement, like autotools etc, but with basic
variables to check if they're loaded.