Add tox env for external standard plugin tests
Adds tooling to define a list of external plugin repos in
`requirements/external-requirements.yml` in a nice yaml format, which
seemed the easiest way to sort out the potential config issues. The main
thing is the addition of the `requirements/parse-external.py` script,
which generates a list of package names and two requirements.txt files,
one for fixed versions, one for master.

It is assumed that `master` more means "test against a branch" and
`fixed` means "test against a tag".

Also adds tox environments to run these tests.