loader.py: Avoid loading deps of junction metaelements
By avoiding this, loading metaelements of junctions becomes cheap even
for junctions with erroneous dependencies, and we can ignore their
task reporting.

Task reporting for junction metaelement loading is confusing, since
the junction element itself will never be part of the pipeline, so
we'd rather not have this show up as an actual loaded element.

Elements loaded from the junction are loaded separately, therefore
this does not affect their progress display.