testutils/context.py: Mock tasks instead of accepting Nones
To ensure that we only disable element loading task progress reporting
for very specific code paths, we need to teach the test suite to be a
bit smarter.

For this reason we now mock a _Task object and return it in our mock
context's relevant method invocations.

Other code paths that deliberately invoke the loader without task
reporting now mark their loads with NO_PROGRESS.