Remove XXX comment about missing progress
This should be safe now - this particular point turned out to be
involved in loading dependencies of junction elements, rather than
anything in their projects.

This meant that, yes, we were missing progress, however junction
elements are not allowed to have dependencies in the first place, so
we simply short-circuit their load and avoid the problem altogether.

We also added more explicit progress opt-outs, since it's far too easy
to end up with spurious Nones.