Add external plugin tests to CI
Runs external plugin tests on fedora 29 with allowed failure.

At present this only tests against a branch, as there is tooling missing
in bst-plugins-experimental, and there is no tag which would be
compatible. In future there should be two targets, one testing plugins
master and one testing the last tag.