tests/frontend/buildcheckout.py: Create required symlinks on the fly
Windows checks out symlinks as a file with content is the path of the
symlink. This makes it impossible to mount the buildstream folder in
a docker container and run the tests.

This removes the symlink, and creates it on the fly for tests that use
it, which fixes the problem.