cascache: extract CASDProcess in new module
Make it easier to specialize handling of the buildbox-casd process on
Windows, by splitting it into it's own class. This allows us to
encapsulate some decisions, and decreases the complexity of the CASCache
class.

Take some of the complexity out of this file by splitting the
responsibility of managing the process out to another file.