_fuse/mount.py: Monitor the fuse process while waiting for the mount
In some cases, users might not have permissions to use fuse, or fuse
might crash.

This was previously leading to a hanged process and, with chance an
error message on the UI, which could be overwritten.

This ensures we are explicitely monitoring the fuse process while
waiting and adds better reporting of the fuse errors.