.gitlab-ci.yml: Assert that tests without `usedevelop` work
This is partially to prevent issues like #1121, but also to ensure our
manifests can be packaged correctly.

For now only the test suite is exercised, in the future it would be
nice to expand this to other tests (such as the man pages, or source
tests), but those don't have -nocover versions yet.