Extract casd_channel logic to CASDChannel
Encapsulate the management of a connection to CASD, so we can hide the
details of how it happens. This will make it easier to port to Windows,
as we will have to take a different approach there.

Remove the _LimitedCASDProcessManagerProxy, as CASDChannel does
everything the child jobs will need.

Also make get_local_cas() public, since it is already used outside of
the CASCache class. Make get_cas() public to match.