casserver.py: Run buildbox-casd in `bst-artifact-server`
This is in preparation of a switch to directly calling buildbox-casd
through the `bst-artifact-server`, rather than relying on BuildStream
internals to do this.

This should help untangle the codebase a little, since our CAS
interfaces will all end up with a single goal, rather than trying to
do both server-end and client-end things.