cli: BST_FORCE_START_METHOD only sets if necessary
Allow situations where the start method is already set, this enables us
to use this in testing situations.

Also, print a diagnostic if it's already set to something we didn't
want.

Now this block got more complex, split out into a new function.

Now we're using this string a lot, extract it to a variable, to make
sure we're spelling it correctly everywhere.