testing, messenger: make dummy_context picklable
Note that unittest.MagicMock is unfortunately not pickable, so if one
tries to cross over to a ChildJob then we're in trouble.

Sacrifice some convenience and a bit of safety by implementing our own
_DummyTask instead of using MagicMock.

Don't try to pickle the 'simple_task' context manager, if it's been
overridden when creating a test context.