_plugincontext: stable ids for better pickling
Use the 'identifier' argument of PluginBase's make_plugin_source(), to
ensure that the plugins have a consistent namespace when pickled and
then unpickled in another process.

This means that we can pickle plugins that are more than an entrypoint,
e.g. ones that have multiple classes in the same file.

This enables the `tests/frontend/mirror.py::test_mirror_fetch_multi`
test to pass, which uses `fetch_source` plugin. That plugin has more
than one class in it's file.