tests/sources: server.start() before using port
Re-order the calls of `server.start()` and
`generate_project_file_server()` such that we are not relying on knowing
the server's port before we've called 'start' on it.

This allows us to use spawning or forking to start the server process,
as we won't be relying on being able to share un-picklable things
between the processes.