testutils/artifactshare: don't hang on error
Remove a couple of cases where it's possible to make the main test
process hang, waiting for something to appear on a queue.

Make it clearer what is available to the server process by
splitting it out into a non-member function.

Raise a friendlier exception, earlier, if there was a problem starting
the server process.