[WIP]Split dependencies into dependency-names and dependency-keys
In Element._calculate_cache_key the dependency field of the cache key
dictionary is replaced with the dependency-names field and the
dependency-keys field. The first is only populated in the case where the
weak dependencies are passed and the second is only populated when the
strong dependencies are passed. Additionally, in the case where the
element holds at least one source advertising BST_NO_REF, the new fields
are both initialised to `[]` in order to correctly update dependencies.
Failing to reinitialise dependency fields in this case will lead to
erroneous dependencies from old states of BST_NO_REF sources.