cascache.py: Set up socket path via a symlink
This is necessary to allow using buildbox-run with userchroot in the
near future, since currently only the owner of the BuildStream process
can access the CASD socket, but the buildbox-casd binary will need to
be setuid' to another user.

This gets around this limitation by allowing the group to access a
symlink, which in turn should point to a directory owned by the CASD
user.