Make temporary staging directories group-accessible
This is again required because of the userchroot workflow requiring
two UIDs. By default, python will create temporary directories with
u+rwx,og-rwx, so that we don't leak information to the public /tmp.

This is irrelevant here since we are in directories that shouldn't be
as easily accessible anyway (since they are usually far inside
"$HOME", or at least configured by the sysadmin to be reasonably
safe), and a hindrance to adopt userchroot as a first-class sandbox.