Add _is_trackable() method to Source()
This method reports whether the source can be tracked. This would be
false for sources advertising BST_KEY_REQUIRES_STAGE.

Element tracking can be skipped if none of the held sources can be
tracked. This is determined by the value of the
`Element.__tracking_scheduled` attribute which is set in
`Element._schedule_tracking()`. This is set to `True` if at least one
source can be tracked.

Also remove some of the tracking handling from `_stream._load` to
`_stream.track` where it is more relevant.

closes #1186