scheduler.py: Prevent the asyncio loop from leaking into subprocesses
Having a running asyncio loop while forking a program is not supported
in python and doesn't work as expected.

This leads to file descriptors leaking and the subprocesses sharing the
same loop as the parents. This also leads to the parent receiving all
signals the children receive.

This ensures we don't leek our asyncio loop in the workers we fork.