job.py: Don't use 'terminate_wait', as it uses waitpid()
Using `join()` on the subprocess calls `waitpid()` under the hood
which breaks our child watcher.

Instead, schedule a task for 20 seconds later that will effectively
kill the tasks.

Note that the task will only be called if we still have active jobs.
Otherwise, it will just be skipped and we won't wait as long.