tests/frontend/completions: Fix implicit string concatenation
We expect `INVALID_ELEMENTS` to be a list of element names. However we
are missing a comma in between the names, so they end up being
concatenated into a single strings.

The test passes either way (which is not ideal) because it just checks
that the suggested completion is _different_ than the expected in case
of errors.