_frontend: Simplify color handling
Currently we store color configuration in our App object, and have to
remember to pass it around to `click.echo()` when printing things on the
screen. This is error-prone as we can forget to do so. This leads to
bugs like #1200, where `bst init` was not respecting `--no-colors` flag.

Instead of doing that, this patch controls colors in output by
configuring the `click.Context` object.

Fixes #1200.