plugins/sources/deb: Re-enable import check for arpy
In the olden days on linting BuildStream, we did not had plugin
dependencies installed in the environment. Now that we run these things
via tox, we have all the dependencies (including plugin dependencies) in
the virtual environment that tox uses to run the commands.

So remove the disabling of import errors for `arpy` module.