element.py: Compute whether an element is cached only on `is_cached`
This removes the early call to get whether sources are locally cached
in `_update_source_state` by delegating it to the call of `is_cached`.

Once it is cached, the element is assumed to stay that way for the
duration of the pipeline, we can therefore cache the result once it is
true.

Also remove `Consistency.IS_CACHED`, which is not used anywhere else.