optionpool.py: Make jinja autoescape rules explicit
Our security linter warns us that jinja should be set to escape HTML
output. Since we don't do HTML output, or any other markup that would
be vulnerable to XSS, we explicitly disable it on our strings, and let
jinja do as it pleases on files.

Should anyone use BuildStream as a library, they should likely escape
our output before displaying it in a browser, but that's a given since
they are operating on user-defined data.