scheduler.py: Optimize scheduling by not calling it unnecessarily
This delays the call to the re-scheduling of jobs until the current
event loop as terminated.

This is in order to reduce the number of time we call this method per
loop, which should reduce the pressure on the loop and allow faster
event handling

Since the call is now delayed, also ensure we only call it once per loop
iteration.