_project.py: Allow junctions to use parent remote
At present it doesn't seem to be possible to use ignore-remote-caches
and also cache cross-junction artifacts in one's own cache. By passing
the parent caches to the junction we ensure that things get cached in
the parent cache.

For a motivating purpose, consider that one may have a (patched)
junction which specifies a cache incompatible with master. This will
throw warnings at every invokation of bst, or you won't cache
cross-junction artifacts.