_gitsourcebase.py: Manage submodules recursively
Previously, GitSourceBase would only consider immediate submodules of the
superproject. It now fetches and stages recursively.

To achieve this, this commit somewhat refactors the relationship between
GitMirror and GitSourceBase. Enumerating GitMirrors for the submodules is now
done in GitMirror itself. GitSourceBase recursively iterates these mirror
classes with _recurse_submodules and applies the source configuration with
_configure_submodules.