utils.py: Use `onerror` in `_force_rmtree`
If we don't, and encounter a file we don't own, but have permission to
delete, we'll fail with EPERM, since we won't be able to change
permissions but will be able to delete it.

Instead, we now try to change permissions and remove a file *after* we
realize we couldn't at first.