_gitsourcebase.py: Refactor fetching and tracking
This changes how the Git repo in the source cache is managed.

The mirror is not created with an initial clone, but is set up manually from
an empty repo. This is to support shallow cloning in future changes.

To simplifiy code and Git calls, remotes are no longer used; instead URLs
are passed on the command-line.