setup.cfg: Don't collect tests from src directory
Generally we don't have any tests in the `src` directory so we don't
need to collect anything from that directory. The only exception to this
are the `sourcetests`, but they add their own collection hook so they
are not affected by this.

This fixes an issue where pytest gets confused upon finding two
different `conftest.py` modules - one in the `src` directory and one in
the virtual environment.

Fixes #1121.