setup.py: Remove check for bubblewrap
Bubblewrap is not a hard requirement for BuildStream now that it
supports other sandboxing mechanisms, i.e. buildbox-run and remote
execution.

Aside from that, this check is already a bit misplaced in `setup.py`.
Bubblewrap is a runtime requirement, so warning about it at build time
isn't quite right. It makes sense when BuildStream is being installed
from source, but not when it is installed from any binary distribution.

Note that the corresponding sandbox implementation already does this
check at runtime.