_gitsourcebase.py: Strip git-describe tag info from cache key
The beginning parts of git-describe labels are completely arbitrary.
They can be changed either manually or by a track (e.g. if a tag is
moved, added or deleted) even if the referenced commit is the same.
Hence, only the commit ID part of the label should factor into the cache
key.

This commit, of course, breaks cache keys for artifacts built with the
'git' source with git-describe refs.