Plugins: Update entrypoint group for pip origin
As discussed in
https://mail.gnome.org/archives/buildstream-list/2020-January/msg00001.html,
separate the entry point groups for elements and sources. Previously,
both used `buildstream.plugins` which meant that it wasn't possible to
have an element and a source plugin with the same name, when using `pip`
origin.

This a breaking change for all plugins, but will only affect projects
using the `pip` plugin origin.

As part of this, also bump the version of bst-plugins-experimental used
in our tests to avoid CI breakages.