_sandboxbwrap.py: Create /dev/shm in the sandbox
Create /dev/shm as a tmpfs in the sandbox. Before now access to /dev/shm
was only available by a plugin using `Sandbox.mark_directory()` or
adding to `Sandbox.DEVICES`, either of which would _mount_ /dev/shm into
the sandbox, allowing pollution from the host. This adds it as a tmpfs
by default, which seems sensible as it is required for POSIX support.

Also adds a test which makes sure that we can open a shared memory
object inside the build sandbox with some (probably poor) C code.

Backport of !1694