Drop concept of partially supported sandbox configuration
Allowing builds without affecting the cache key but disallowing push,
when the sandbox configuration is not fully compatible, results in an
inconsistent user experience and may lead to unexpected build issues.
Especially as push is allowed or disallowed based on the sandbox
available at the time of push, not at the time of build.

The previous commit making build-uid and build-gid configuration
optional allows each project to decide whether the sandbox is required
to support UID/GID configuration or not.