pip source plugin: Add support for python 3.8
For some reason, the pip source plugin appears to pass tests in master
when testing python 3.8 without this change, however it does not pass
tests without this change on the bst-1 branch.

Adding this change to master for consistency across branches for now,
and in the off chance that this fixes the pip source plugin in some
python 3.8 environments.