doc/source/using_tutorial.rst: Renamed to "Getting started"
This section of the user manual describes the basics of creating
your first BuildStream project, while we've discussed this in
terms of a "Getting started" guide while developing it, we ended
up naming this a "Tutorial" because of it's walkthrough nature.

Due to it's name as a Tutorial, developers (our target audience)
have a tendency to avoid it and look for more terse and advanced
material, while this material could be more suitable for getting
started.