_artifact.py: Don't consider an artifact cached if logs are missing
Artifact push and pull operations currently fail if logs are missing.

We don't currently have a config option to control how long artifact
logs should be kept in the cache. Until this changes, we should be
conservative and consider logs to be an essential part of artifacts,
keeping them from getting expired before the rest of the artifact.