CONTRIBUTING: Remove link for obsolete issue policies
As discussed in
https://mail.gnome.org/archives/buildstream-list/2020-April/msg00009.html,
we want to get rid of the nosoftware subgroup.

These issue policies are the only remaining link from here to the
nosoftware subgroup.

I propose to remove this link because the linked are obsolete and overly
process-oriented. Our current practices don't reflect those policies
either. For example, I don't think any of us are religiously using
labels like backlog/todo/doing etc.

In my opinion, the description at
https://gitlab.com/BuildStream/buildstream/-/labels is sufficient for
documenting what labels mean, and we don't need to force everything to
have one label of each category.

We can obviously improve the label descriptions if they are lacking.