tests/plugins/loading.py: New test replaces removed internal test
This test tests some of the basic failure modes of plugin loading,
which used to be written as internal tests but is not implemented
as a proper end-to-end test.

This commit also adds some machine readable reason codes to
pluginfactory.py so that we can assert the errors more specifically.