plugin.py/pluginfactory.py: Implementing BST_MIN_VERSION
The BST_MIN_VERSION guards assert that the BuildStream core which
loaded the plugin is compatible with the plugin itself.

This commit adds BST_MIN_VERSION to the base plugin.py with documentation
informing Plugin authors how to set the minimum version, and also adds
the assertions at plugin loading time in pluginfactory.py.

This commit also:

  * Adds the BST_MIN_VERSION specification to all current core plugins
  * Adds the BST_MIN_VERSION specification to plugins used in test cases